Fix confusing sorting in data handling
After running sqldf on data for the 'file' artifact, the data got re-
sorted, which made data comprehensibility more difficult. The data is
now sorted the same way as in the codeface-extraction project.

Additionally, the data in the thing2thing method is now kept in the same
order as in the input data. This will also increase traceability of
data.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>