Add artifact2author data method
In order to get the list of developers working on an artifact, the new
method 'artifact2author' can be used. The base artifact is filtered by
default.

The commits for an artifact are sorted by date, so that a directed
network is possible to build based on the data given by the method.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>