Add date offset to mail data
As the codeface-extraction project added timezone offsets for mail data,
this framework now incorporates the data, too.

IMPORTANT: This is potentially a breaking change, as working with older
data (non incorporating the tz data) shifts columns, so that columns in
the mail data of this framework get shifted to different columns!
Update your copy of the Codeface data!

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>