Parallelize construction of networks
To reduce time when constructing networks, some code parts during the
construction of a single network are now parallelized using 'mclapply'.

The construction of several networks using the bulk functions is not
parallelized, since the over-parallelization would reduce the speedup
again.

For the configuration 'busybox-function-cochange', the time for
constructing a project-level author network is reduced from over three
days to 10 seconds on my machine.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>