Partly revert "Parallelize construction of networks"
This partly reverts commit 91cb0aaa6cc5f88ccdc782a8c58837ae8d622c72.

As the parallelization of the edge construction involves quite an amount
of memory for (de-)serialization is needed and this amount is restricted
to 2GB [1,2], the parallelization breaks for directed artifact-based
networks on project level.

This fixes #3 by the cost of partly removing parallelization (at least,
for the moment.)

[1] https://bugs.r-project.org/bugzilla/show_bug.cgi?id=17137
[2] http://stackoverflow.com/q/23231183

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>