Fix error when constructing co-change-based author networks
The argument 'filter.base.artifact' passed to the method
'get.author.network.cochange' has not been passed on to the method
'get.artifact2author'. As the argument was set to FALSE instead, this
resulted in not removing the base artifact from the commit list. By
passing on the argument, the argument gets respected now and the method
does not lead to unexpected results.

A long-term consequence of this error was the introduction of too many
edges in the function 'construct.dependency.network.from.list', which is
the explanation for the run-time behavior reported in issue #3.
Basically, the plain re-introduction of parallelization (issue #4)
should be possible now.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>