Respect directedness of networks more carefully
For bipartite networks, we can construct directed author networks while
sticking to undirected artifact networks. This produces problems when we
union the networks to a combined (bipartite) network.

The current solution is to transform the artifact network to the same
directedness of the author network.

[Comment: Unfortunately, this is not the best solution, but a first
step. I will open an issue on that topic.]

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>