Skip dependency lists which cause many, many edges
To be able to skip the edge construction for things that cause many,
many edges, we introduce a threshold, which will ensure that any things
with an edge amount above a threshold are skipped.

By default, the threshold is "Inf", i.e., there is no threshold.

If any thing is skipped, a warning is issued with detailed information.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>