Add activity-based splitting functionality for Codeface*Data objects
In the file 'util-split.R', we now provide the functionality to split a
given Codeface*Data object according to an amount of activity to be
performed in each range. As a result, a list of CodefaceRangeData
objects is returned, each containing an amount of activity.

In detail, the amount of activity on either mails or commits is
transformed to a timestamp-based splitting by obtaining the date
attributes of the commits inside a range of data. Based on this time-
based bins, the splitting of the other data sets is performed.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>