Apply simplification parameter to artifact networks
The parameter 'simple.networks' was not respected for co-change-based
artifact networks, consequently, all such networks have been simplified
without consent.

Additionally, the reset of the edge weights in the simplifification
process is dropped to ensure compatibility with networks which have been
simplified before. The initial edge weight of 1 is set when the networks
are constructed (see 'construct.dependency.network.from.list').

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>