Add activity-based splitting functionality for networks
In the file 'util-split.R', we now provide the functionality to split a
given network according to a given amount of edges. As a result, a list
of networks is returned, each representing activity of the specified
amount.

The activity can be specified by the amount of edges per range or via
the number of ranges to be created.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>