Improve the time-based splitting of dates
In the function 'split.get.bins.time.based', we used to use 'split' for
the splitting of the dates vector. The consequence of this step is that
the split bins are starting each at the first day of the month/week/...

By using 'findInterval' instead, we are able to create bins starting at
any date.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>