Add setter for revisions to project configuration
When splitting data objects using the split functions, the project
configuration is updated with the revisions, ranges and revision dates
obtained from the splitting data. This way, retrieving the ranges from
any configuraton object gets the right ranges.

Thanks to the object orientation, the original configuration that was
used to construct the to-be-split data object is also updated. This way,
also the original configuration object (when stored in a variable) is
updated.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>