Add setter for revisions to project configuration
When splitting data objects using the split functions, the project
configuration is updated with the revisions, ranges and revision dates
obtained from the splitting data. This way, retrieving the ranges from
any configuraton object gets the right ranges.

Thanks to the object orientation, the original configuration that was
used to construct the to-be-split data object is also updated. This way,
also the original configuration object (when stored in a variable) is
updated.

This is a backport of commit 9f7219f73a86f5cc4a33afa206a122ef07338b3c
for version 1.0.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>