Fix wrong calls to project and network configuration
Inside the data classes, the network configurations is often referenced
without the 'private$' prefix. This fails if the original network
configuration is not stored in a variable called 'network.conf'. Thus,
the prefix is now appended.

Additionally, there is one call to the late 'get.conf()' method of the
data objects during splitting. This is corrected by this patch.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>