Do not always reset environment when updating project configuration
When updating the project configuration, it may not always be a good
idea to reset the data object. E.g., when splitting a data object, the
ranges in the project configuration get updated and, thus, re-reading
the data from disk is not possible.

This basically adds to 98ae7f08a8fecbe40674f0f00bae98b81769c6d8.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>