Fix time-based network splitting
Passing too many arguments to the logger is a bad idea, as it would stop
the script. Also, wrong logging messages are also problematic.

Additionally, in the time-based network splitting, with the absence of a
default value for the time-period argument, the splitting by bins may
cause problems when giving no 'time.period' parameter. This is fixed
with this commit.

This is a backport of ba6ba6505285e5120ddb9495a05b6519dfcac72f.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>