Fix problem with activity-based network splitting
When splitting networks in an activity-based manner, it may be the case
that the edges for one single event (e.g., bipartite edges when a
developer works on many, many artifacts in a commit) are split into
parts. As a result, the date bins for the networks may contain a pattern
like this '[..., d, d, ...]', where one date is the start of two ranges.
Calling 'unique' then is a very bad idea.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>