Fix bugs in bin retrieval for splitting functionality
There are two problems with the identifaction of split bins, both time-
based and activity-based:
1) When using 'Date' objects for breaks, the identifaction of bins may
break due to unknown reasons.
2) When the designated amount of items (mails, commits, or edges) per
time window is higher than the actual existing number of items, the
identification of bins per item is broken due to '1:0' being a non-empty
vector.

This is a backport of 68e910b2fc4533e6085ca1035ca3cd63c5f2b670.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>