Fix offsets in activity-based sliding-window preparation
When cutting the data/network for the sliding-window time ranges, we
need to round both the start and end offset, as both values can be
uneven.

After some calculations, we decided to use 'floor()' to round the
values. It may be better to have more data available than less (although
it's about 1 or 2 data items).

Props to @bockthom for his help.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>