Respect 'contract.edges' when constructing multi network
When the author network is undirected, but artifact network is not, then
the union for the multi network does not work. To this end, the artifact
network is converted to an undirected one. Doing this, edges can be
contracted or not---as configured by the parameter 'contract.edges' in
the network configuration.

Due to a mistake, the parameter was not respected earlier (see
88c5e26fc4c72b96848c7d8e7dd6a05667cb46fc). The error affects versions
v1.0 and v2.0.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>