Add edge and vertex attributes earlier in construction process
All types of networks need to have proper edge and vertex attributes.
Before, the attributes have only been added for bipartite and multi
networks.

To introduce edge and vertex attributes for all types of networks, the
attributes are now added early in the 'get.*.network' methods of the
'CodefaceProjectData' class. In the same process, the control-flow
mechanism is altered slightly to add the attributes properly.

With the new implementation, the attributes do not have to be added for
multi networks anymore. Additionally, all 'plot*.network' functions work
as expected.

This fixes #32. Props to @bockthom for pointing this out.