Fix edge-count calculation during edge construction
The calculation for the amount of directed edges to construct in one set
had an offset by one. This is corrected in this commit.

This error could have affected the 'skip.threshold' parameter of the
network configuration.

This is a backport of 7a1bf4e215e6dc2238ece1aa0c379d975d101149.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>