Refactor motif modification to handle single vertex sequences
Since now, the function 'motifs.remove.artifacts.from.matched.motifs'
could only handle a list of vertex sequences 'vs' as parameter. This is
changed such that also a single vertex sequence can be passed.

The result of the function (independent of the 'vs' argument) is always
a list of vertex sequences, all stripped from artifact vertices.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>