Fix problem with empty artifacts appearing in networks
For co-change-based artifact networks, the relation commit--artifact is
used as a basis to generate edges. We do not want the empty artifact
(i.e., "") to appear in the networks, therefore, we use the properly
filtered commit list as an input.

Analogously for the relation commit--file.

Props to @bockthom for the debugging session.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>