Add parameter 'author.all.authors' to network configuration
With this patch, the new network-configuration parameter
'author.all.authors' is introduced, which denotes whether all available
authors (from all analyses and data sources) shall be added to the
network as a basis. Depending on the chosen author relation, there may
be isolates then.

The parameter 'author.only.committers' is now a complement for the
parameter 'author.all.authors' and removes in a second step all authors
that do not commit to the repository.

The code is cleaned up and tests for this specific kind of network
construction are introduced.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>