Fix problem with wrong bins during time-based splitting
When splitting data or networks in a time-based manner, it may happen
that some ranges do not have any data or edges. In that case, the empty
ranges had been ignored and filled with data.

The exact problem was the assignment of the date bins for the bin vector
in 'split.get.bins.time.based'. By carefully setting the factor's levels
to the right values, the problem is fixed.

Additionally, some clean-ups in this function are performed.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>