Add test on consistency of data and network time-based splitting
The chosen time-window size during splitting of data objects may result
in the following condition (described on a co-change-based author
network):
1) Person A only appears in the first time window, working on artifact X
as the only author.
2) Person B only appears in the second time window, working on the very
artifact X as well.
Thus, when splitting the project-level network, there are edges from
Person A to Person B, crossing the time-window border, which do not
appear when constructing the author networks on the split project data.

Note: We reached 100 test cases now!

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>