Fix bugs in sliding-window splitting of data objects
The data subset on the basis data in the acitivity-based data splitting
works properly only when performing a subset on the rows, not the
columns (comma was missing). This is fixed.

When using the sliding-window approach for data splitting, the number of
"normal" ranges is relevant: If we have only one normal range, the
sliding-window approach does not make sense. In this case, a warning is
now issued and the sliding-window approach is not applied.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>