Fix and refactor 'get.key.to.value.from.df'
The function 'get.key.to.value.from.df' needs to return a data.frame
with 'thing2' as the first column in each list item for the function
'construct.dependency.network.from.list' to work properly, as 'thing2'
is used as the vertex names in the network to be constructed. This is
fixed.

Additionally, no columns are dropped anymore, i.e., the columns
indicated by 'thing1' and 'thing2' are additionally inserted at the
start of the data.frame. This way, we are able to support all columns as
edge attributes later.

The documentation is extended significantly and the tracing is slightly
changed.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>