Restructure package dependencies in .travis.yml
It is possibly enough to install the packages through the file
'install.R', so that the YAML key 'r_packages' is not needed. This has
also the positive side effect that we do not need to maintain two lists
of dependencies in the repository.

Later, the dependencies should be listed in the DESCRIPTION file.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>