Provide possibility to handle duplicate range names (#43)
During activity-based splitting of networks, it may happen that there
are generated duplicate range names. As these are not suitable for use
in row names of data.frame, this patch provides the function
'split.unify.range.names' which is able to rename range names. For this,
duplicated range names are suffixed by the pattern ' (#)', where '#'
is a number.

When duplicated ranges are discovered during activity-based splitting of
networks, a warning is generated.

Fixes #43.