Fix problem with missing edge attributes
With the new behavior in the function 'get.key.to.value.from.df'
introduced by d8bfff95d75b9beb2230f89cd741d4e7ba3582ff, the resulting
data.frame had, for example, the two columns 'artifact' and
'artifact.1', where the first one is equivalent to 'thing2'. As a
consequence, the edge attribute 'artifact' was missing later, even if
supplied to the corresponding network-configuration parameter, since it
could not be matched with 'artifact.1'.

Fix indentation to 4 spaces.

This is a fixup for d8bfff95d75b9beb2230f89cd741d4e7ba3582ff.