Remove e-mails with obviously wrong dates (#50)
It may occur that there are e-mails with obviously wrong dates (such as
"107-07-20 16:35:57"), since these date formats are compatible with the
"%Y" pattern. As there can no e-mails be sent before 1971, we now remove
all e-mails with a date before 1970-01-01 00:00:00.

Fixes #50. Thanks to @bockthom for pointing this out.

This is a backport of 91af305e9202cfedeff9b73ec58f668af9a3c207.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>