Add progress to edge-construction logging
To have better track of the progress during edge construction, the
progess is now incorporated in the log message.

The progress is notated in the form "[n/m]" as string prefix to the log
message, where m is the total number of list entries from the input data
and m is the index of the current item.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>