Fix error with commit sorting when enabling synchronicity data
When we add the synchronicity data (or any other data source) to the
commit data (or any other data source), we must not change the ordering
of the row in the respective data.frames. Especially, when using 'merge'
to join data, the 'sort' argument must be set to 'FALSE'.

Props to @bockthom for spotting this.

This is a backport of 59dc66db03557c572bbf5366a2b52bb07fce6d88.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>