Update plotting functionality to 'ggraph' package
For better usability for plotting, we now use the 'ggraph' package,
which builds on 'ggplot2'. This way, we are better able to tailor the
plot to the specific needs (e.g., using facets).

The plotting can be achieved through two functions now:
1) The function 'plot.print.network' (or the alias 'plot.network')
constructs a plot and prints it directly using the 'print' function.
2) The function 'plot.get.plot.for.network' constructs a plot object
that can be manipulated before printing (e.g., add facets, or use
'ggplot2::ggsave' for printing).

The showcase file 'test.R' is updated accordingly.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>