Fix role-stability analysis
There were problems when no developers were absent in the evolution. This has now been fixed by setting the probability of transitioning to core/peripheral from absent to 0. The probability of transitioning from absent to absent was set to 1, as all developers that are absent remain absent in this scenario.