Fix timezone for reading mail timestamps
The timezone is now fixed to "UTC" to enable reading of dates. There was a problem with one mail that was written on 2015-03-29 02:32:01 CET according to the database (This timestamp does not exist due to the transition to CEST). This is only a workaround and should be looked at again.