Fix timezone setting for reading timestamp data
The default timezone is now set to "UTC" to fix errors while reading
dates and timestamp data. By default, R uses the system's default
timezone (which is usually "CE(S)T" in our case), consequently, the
timestamps got read with a default offset (UTC -> CE(S)T) (at least,
consistently), although the data is available and stored in UTC.

The error was eventually noticed while reading the date "2015-03-29
02:32:01". This date does not exist in CE(S)T, as, at exactly this time,
there was the transition from CET to CEST. In UTC, in contrast, the time
exists.

Props to @sofie-kemper for spotting this and implementing a first fix in
commit 25b90741e58daf58f64391d0e0d012712b302437.

This is a backport of 4a3f8b91317fe7953670a83366cb3ad9c74b4d42.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>