Fix encoding setting for reading data
When not using a/the proper encoding when reading the data files from
disk, the data gets broken and the result may be corrupted. This issue
is fixed by setting the locale "en_US.UTF-8" as default before reading
any data.

This is a backport of ea1925b880bbe7b77528c077f6244f3564007330.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Sofie Kemper <kemperso@fim.uni-passau.de>