Add function for string representation of configurations
With the function 'get.configuration.string', we now have functionality
to properly construct a string representation of our configuration
objects. Basically, it constructs a nested list output that can be
passed to the logging utility.

The main difference with the former implementation is that it is more
reliable and does not transform values (such as "TRUE" to "yes").

An example output (truncated):
...
- artifact.directed
- value = FALSE
- edge.attributes
- value = [
      date,
      message.id,
      thread,
      hash,
      file,
      artifact.type,
      artifact,
      issue.id,
      event.name
    ]
- simplify
- value = FALSE
...

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>