Fix problem with empty issue data
The check whether the issue data has been read properly was wrong. The
check is changed for the 'try-error' construct. In the error case, an
empty data.frame is returned -- analogously to the other data sources.

This fixes the splitting tests. Although, there are still no tests for
issue data in there.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>