Fix setter for issue data
When splitting a data object, the splitting data was set in the
RangeData as NULL if there were no issue data for the current range.
Consequently, the data is read from disk again.

To fix this, the issue data is set as empty data.frame if there is not
issue data available (i.e., if it is NULL). This is also consistent with
the other setters.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>