Change bipartite relation to artifact relation
The bipartite relation is now determined by the 'artifact.relation'
attribute of the network configuration. The main reason for this is that
the type of artifact (e.g., file, function, issue) indicated by the
bipartite relation needs to be same as the one indicated by the artifact
relation; otherwise, a malformed network with several types of artifacts
may be the result.
A consequence here is that, for the construction of bipartite networks
to work correctly and to make sense, the set of allowed artifact
relations needs to be a superset of the allowed author relations.

In this relation, also the behavior of 'author.all.authors' and
'author.only.committers' is changed: Only these two attributes of the
network configuration enforce the appearance of authors in the network,
not the available data anymore. Here, we rely on the fact that the list
of all authors is complete.

The tests are changed accordingly. Some TODO notes are inserted for
(minor) open issues.

Thanks to @sofie-kemper, @Roger1995, and @bockthom for their valuable
feedback on these changes.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>