Update author-network tests with issue relation
The tests for 'author.all.authors' and 'author.only.committers' are
augmented with a test for the issue relation. We perform a code clean-up
alongside.

Additionally, we add a new author to the test data who is only available
in issue data for a more complete test case. Fix reading test
accordingly.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>