Refactor week splitting in core-peripheral analysis
For better code unterstanding, the splitting commit data into ranges of
given weeks is refactored to use the functions available from the
splitting functionality in this library.

Hereby, a serious bug with the splitting mechanism is fixed: the
function 'split' needs a vector as long as the data.frame, otherwise the
result gets weird.

Additionally, the ranges do not have spaces in the separator anymore ("
- " before, "-" now).