Rework sample network to use sample project data
Until now, the sample network has been built by manually constructing
the network parts and calling the right functions, such as
'combine.networks'. This is changed to a structure using a proper
project configuration with data and network-builder objects.

The sample data is placed in the folder './sample', which is configured
project-wide by the variable 'SAMPLE.DATA' as an absolute path.

The project configuration is 'ProjectConf$new(SAMPLE.DATA, "testing",
"sample", "feature")', while the network configuration is built as
follows:
> net.conf$update.values(list(
>     author.relation = "mail", author.directed = FALSE,
>     artifact.relation = "callgraph", artifact.directed = FALSE,
>     author.all.authors = TRUE, author.only.committers = FALSE,
>     simplify = TRUE
> ))

For plotting purposes, the resulting network is given a proper layout
attribute.

As the sample-network construction is now using the default network-
construction functionality, the order of the vertices in the networks
and, thus, the test results are mixed up. The tests are fixed
accordingly.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>