Move network-relation functions from misc to networks module
In this patch, many network-relation functions are moved to the networks
module of the library. Due to this, the modules are better modularized
again now.

The function 'construct.dependency.network.from.list' is renamed to
'construct.network.from.list'. Addtionally, the functions
'read.network.from.file' and 'postprocess.artifact.names.callgraph' are
removed entirely.

This patch also fixes the problem of call-graph-based artifact networks
for functions are not properly post-processed (removing obsolote
absolute paths etc.).

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>