Refactor 'motifs.count' and 'motifs.count.all'
For better interoperability and to keep the door open for further kind
of motifs, the methods 'motifs.count' and 'motifs.count.all' are
refactored to handle a list of named motifs, respectively.

The motif-related entries in the returned lists are named by the names
of the motif list passed to the function. This way, we gain a more
generic approach for searching a set of motifs.

The items 'p1' and 'p2' from the returned list are moved to the function
'motifs.count.all' as they are not generically applicable.

The docs and the tests are adapted appropriately.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>