Fix problem with rounding errors in classification
In one case, author.data contained (apparently) two times the value 1, but the cumulative sum was always seen as > 2 (author.class.threshold). This occurred due to rounding errors. This commit contains a workaround, which treats this error case and ensures that the calculations can continue correctly.

This might not treat all possible error cases, only those where the program throws an error. We should discuss this problem.