Fix plots to always respect edge and vertex type
When plotting artifact networks, the vertex shape has been circles
instead if squares; and when plotting bipartite networks, the line type
of edges has been dashed instead of solid.

This is now fixed by introducing named vectors for the vertex and edge
types which are respected by ggraph and ggplot2.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>