Fix problems with the plotting of empty/edgeless networks
When plotting empty of edgeless networks, the subsetting for ill-formed
edge and vertex types do harm by throwing errors during execution. To
fix the problem, we check for the amount of edges and vertices in the
network before accessing the ill-formed types.

This is a backport of 7c0462b3af9656202512c60b81d2e34832f133fc.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reported-by: Barbara Eckl <ecklbarb@fim.uni-passau.de>
Reported-by: Thomas Bock <bockthom@fim.uni-passau.de>