Add range attribute to built networks
When building networks, we sometimes need to identify the network's
revision range properly. To make this possible without keeping a list of
networks and according names of the list, we now add an attribute
'range' to the network -- if and when it is built from a RangeData
object.

This fixes #62.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reported-by: Thomas Bock <bockthom@fim.uni-passau.de>