Add a changelog file
The file 'NEWS.md' now holds the changelog of the network library for
better overview. The main sections are "added", "changed/improved", and
"fixes".

Added: Indicating completely new functionality, such as added features
or functions (e.g., adding the splitting functionality).
Changed/improved: Indicating adapted or updated functionality (e.g.,
updating the plotting functionality with the help of another R package).
Fixed: Indicating real fixes for bugs and regressions (e.g., fixing a
typo in a function call).

Later, the upcoming sections for new versions should be added, on the
one side, automatically generated from the commit list and, on the other
side, with a self-written text, pointing out the overall synopsis and
the highlights.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>