Fix dependency on 'ggraph' package
The dependency on the R package 'ggraph' (and also 'ggplot2') is now
documented in the README file. In particular, the system dependency on
the 'udunits2' system library is included.

The dependency is now properly installed during CI runs. We do not test
for plotting, but running the complete installation routine is a test
itself.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>