Fix date parsing for splitting module and range-bounds parsing
When parsing a vector of timestamps using `as.POSIXct` where the hours
are missing from, at least, one timestamp (indicating midnight), the
hours are stripped from *all* timestamps! To circumvent the problem, we
now parse timestamps given as character strings using the function
`lubridate::ymd_hms(..., truncated = 3)`, which adapts appropriately
according to the given timestamps.

This change is, most importantly, relevant for the parsing of range
bounds from a given range identifier (e.g., "2017-03-04
12:45:38-2018-05-27 12:45:38").

Note that the range identifiers now *always* include a substring
indicating the hours of the timestamps, even when the time is midnight!

Example: `str(as.POSIXct(c("2017-01-02", "2017-03-04 12:45:38")))`
yields `POSIXct[1:2], format: "2017-01-02" "2017-03-04"`, which is
clearly wrong.

Adapt tests appropriately to also test the changed functionality.

Update README, showcase file and install.R on the way.

This fixes the problem in PR #66.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>