Remove system-variable initialization from module files
As discussed in issue #68, we now enforce that the user needs to source
the file `util-init.R` to initialize the network library. The
independent inclusion of individual files is not intended anymore and is
now unsupported.

Respective code in the reading and splitting modules are removed. The
documentation is updated accordingly.

This fixes #68.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>