Fix small bugs in network metrics
In this patch, we fix two minor bugs in the network metrics
'node.degree' and 'modularity'. In the first, the unsorted result was
not assigned correctly to the return value. In the latter, the single
modularity value does not need a name (which was also an undefined
variable).

Props to @ecklbarb for reporting these two mistakes.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>