Add committer data to commit data source
After the landing of se-passau/codeface-extraction#6, we can add
committer data to the resulting data.frame of the commit data source.
This includes "committer.date", "committer.name", and "committer.email",
where the committer date is consequently converted to POSIXct.

The data is also available as edge attribute through the network
configuration.

When reading data from disk and no committer data is available in the
original data, we adaptively add NAs for it.

This fixes #35. This change also needs to be backported to v3.0.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>