Remove warning when reading non-existent PaStA data
We should only give warning about non-existent data for splittable data
(i.e., commits, mails, issues), but not for others (i.e., PaStA and
synchronicity data). This way, the user does not get confused by the
warning, especially, if the data is only attempted to read during
splitting, but is not really used.

Thanks to @bockthom and originally @Angelika1 for reporting this.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>