Add author-classification attributes based on classification result
The function 'add.vertex.attribute.author.role' now adds author-
classification attributes based on classification results. The function
'add.vertex.attribute.author.role.function' is now the one using a
classification function as a parameter.

For this to work, we add the range identifier as a parameter to the
anonymous calculation functions for adding the vertex attributes.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>