Minor fixes from review in PR #93
This patch introduces only some minor fixes and inprovements to the
documentation for the changes in PR #93.

The most important change is the renaming of the function
'add.vertex.attribute.commit.count' to
'add.vertex.attribute.commit.count.helper', mainly for unambiguity and
consistency reasons.

Props to @bockthom for being that precise.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>