Implement function to construct overlapping ranges
As an request from issue #61, we add the function
'construct.overlapping.ranges' in this patch, with which reliably
overlapping ranges can be constructed. the user needs to pass a start
date, an end date, a time period for the ranges, and an overlap time
period. The ranges can be requested in raw format as for
'construct.ranges' if needed.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>