Implement function to construct cumulative ranges
As an request from issue #61, we add the function
'construct.cumulative.ranges' in this patch, with which reliably
cumulative ranges can be constructed. The user needs to pass a start
date, an end date, and a time period for the ranges. Each range will
then start at the start date, but end at different dates, in a
cumulative manner.
The ranges can be requested in raw format as for 'construct.ranges' if
needed.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>