Implement function to construct consecutive ranges
In this patch, we add the function 'construct.consecutive.ranges', with
which consecutive ranges can be constructed (with no overlap). The user
needs to pass a start date, an end date, and a time period for the
ranges.
The ranges can be requested in raw format as for 'construct.ranges' if
needed.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>