Refactor 'split.get.bins.time.based' to use 'generate.date.sequence'
For interoperability and reliability, we now use the function
'generate.date.sequence' to generate the bin bounds that are needed to
split any data in a time-based manner in 'split.get.bins.time.based'.

As, with lubridate, a year is now exactly 1 year and 6 hours to mitigate
leap years, the tests are adapted accordingly. Same thing holds for
months, which are 30 days, 10 hours, and 30 minutes long, on overage.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>