Fix test for 'split.data.by.networks'
As the project end date may come from the timestamp data in a
ProjectData instance, but the end date in any range is *exclusive*, we
need to add one second (1s) when using the aggregation level "complete".
The function was implemented correctly already, but the test was not
adapted to it.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>