Enforce coding conventions
The code is updated to better reflect the coding conventions modelled in
the contribution guidelines. (The only exception is the module 'core-
peripheral', which has issue #70 still pending and, thus, got a TODO
item on this.)

The main items that are being enforced:
- spaces after commas,
- spaces around equals signs,
- equals signs as assignment,
- list access using bracket notation,
- spaces before opening brackets, and
- double quotes instead of single quotes.

Note: In this patch, certainly not all issues have been covered,
especially, proper documentation, so we need to cover that in upcoming
patches. But it's a start.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>