Improve 'add.vertex.attribute.author.role' function
When adding already exisiting author-role classifications using the
'add.vertex.attribute.author.role' function, we do not need the project data,
and also no aggregation level is used. Hence, the corresponding parameters
are removed now. As no data has to be split any more, this also increases
the performance of this function drastically.

This addresses #102 and fixes Issue 1 therein.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>