Fix activity-based splitting with sliding window of data objects
When constructing the sliding-window ranges in activity-based splitting
of data objects, we cut the half of the wanted activity per range from
the first and the very last range, then split the data. In the
following, we need to set the splittable data properly again.

With this patch, we fix the setting of the cut issue data.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>