Introduce function 'match.arg.or.default'
Add a new function 'match.arg.or.default' for argument verification which is able to take a specified default value.

As the 'match.arg' function of R-base takes the first element of the choices vector as default, we introduce a new function to allow specifing another default value. If no default value is specified, the function 'match.arg.or.default' just calls the 'match.arg' function of R-base.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>