Change default plotting layout to 'igraph::layout.kamada.kawai'
When plotting networks using the default layout of the package 'ggraph'
(i.e., 'igraph::layout_nicely'), we experienced segfaults in R.
Therefore, we change the default layout to
'igraph::layout.kamada.kawai'.

This essentially fixes #109.

Thanks to @hechtlC for experiencing the problems and @bockthom for
helping to find a solution.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>