Change/Improve values of vertex-type and edge-type constants
The vertex and edge types as defined in the networks module
('TYPE.AUTHOR', 'TYPE.ARTIFACT', 'TYPE.EDGES.INTRA', and
'TYPE.EDGES.INTER') are defined as magic numbers (historically coming
from vector indices). We refactor them to be of type 'character'. This
way, we are able to actually "read" the vertex and edge attributes
without any lookup from the network objects we construct.

The constants are now defined as follows:
TYPE.AUTHOR = "Author"
TYPE.ARTIFACT = "Artifact"
TYPE.EDGES.INTRA = "Unipartite"
TYPE.EDGES.INTER = "Bipartite"

As the motif identification works on numerics for the vertex encoding
(see 'igraph::subgraph_isomorphisms', method 'vf2'), the corresponding
code is adapted to transform the (now) character constants reliably back
to numerics. For this, we introduce the mapping 'MOTIF.TYPE.MAPPING' and
the function 'get.vertex.types.as.numeric'.

As we manually translate vertex and edge types to strings for display in
the plotting module, the module gets appropriate fixes and clean-ups.
Here, the legend name for author vertices is not 'TYPE.AUTHOR', but
"Developer". The behavior can be overwritten via setting
'PLOT.VERTEX.TYPE.AUTHOR' (and 'PLOT.VERTEX.TYPE.ARTIFACT',
respectively) appropriately.

This fixes #110.
Thanks to @bockthom for his numerous remarks on this change.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>